refactor: rename `token_pooling` module to `compression`
