+++ colpali_engine/compression/__init__.py
@@ -1,0 +1,1 @@
+from .token_pooling import BaseTokenPooler, HierarchicalTokenPooler, TokenPoolingOutput
